pyproject.toml
[tool.poetry]
name = "qask-backend"
version = "0.1.0"
description = ""
authors = ["chen <23054115+cwkang1998@users.noreply.github.com>"]

[tool.poetry.dependencies]
python = "^3.9"
Flask = "^1.1.2"
Flask-Cors = "^3.0.10"
Flask-PyMongo = "^2.3.0"
greenlet = "^1.0.0"
Flask-SocketIO = "^5.0.1"
eventlet = "^0.30.2"
isodate = "^0.6.0"
hashids = "^1.3.1"
python-socketio = "^5.4.0"

[tool.poetry.dev-dependencies]
black = "^20.8b1"
pylint = "^2.7.4"
mypy = "^0.910"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
